requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
assertpy==1.1
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
exceptiongroup==1.2.2
execnet==2.1.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-xdist==3.6.1
requests==2.32.3
requests-to-curl==1.1.0
tomli==2.2.1
urllib3==2.3.0
